Autism data

Data from a prospective longitudinal study following 214 children between the ages of 2 and 13 who were diagnosed with either autism spectrum disorder or non-spectrum developmental delays at age 2.

data(autism)

Format

A data frame with 604 observation on the following 7 variables:

childid

Child ID.

sicdegp

Sequenced Inventory of Communication Development group (an assessment of expressive language development) - a factor. Levels are low, med, and high.

age2

Age (in years) centered around age 2 (age at diagnosis).

vsae

Vineland Socialization Age Equivalent

gender

Child's gender - a factor. Levels are male and female.

race

Child's race - a factor. Levels are white and nonwhite.

bestest2

Diagnosis at age 2 - a factor. Levels are autism and pdd (pervasive developmental disorder).

Source

http://www-personal.umich.edu/~kwelch/

References

Anderson, D. K., Lord, C., Risi, S., DiLavore, P. S., Shulman, C., Thurm, A., et al. (2007). Patterns of growth in verbal abilities among children with autism spectrum disorder. Journal of Consulting and Clinical Psychology, 75(4), 594--604.

Anderson, D. K., Oti, R. S., Lord, C., & Welch, K. (2009). Patterns of Growth in Adaptive Social Abilities Among Children with Autism Spectrum Disorders. Journal of Abnormal Child Psychology, 37(7), 1019--1034.