Calculating influence diagnostics for HLMs.

This group of functions is used to compute deletion diagnostics for a hierarchical linear model based on the building blocks returned by case_delete.

diagnostics(object)

# S3 method for case_delete
cooks.distance(model, ...)

# S3 method for case_delete
mdffits(object, ...)

# S3 method for case_delete
covtrace(object, ...)

# S3 method for case_delete
covratio(object, ...)

# S3 method for case_delete
rvc(object, ...)

Arguments

object
an object containing the output returned by case_delete()

model
an object containing the output returned by case_delete(). This is only named differently to agree with the generic.

...
do not use

Details

The primary function is diagnostics which returns either a list or data frame of influence measures depending on whether type = "both" (list) or if only one aspect of the model is selected (data.frame). If type = "both", then a list with Cook's distance, MDFFITS, COVTRACE, and COVRATIO are returned for the fixed effects and relative variance change (RVC) is returned for the variance components.

The methods cooks.distance, mdffits, covtrace, covratio, and rvc can be used for direct computation of the corresponding diagnostic quantities from an object of class case_delete.

Note

The results provided by this function will give exact values of the diagnostics; however, these are computationally very slow. Approximate versions of cooks.distance, mdffits, covtrace, covratio are implemented in HLMdiag, and can be called directly on the mer object.

References

Christensen, R., Pearson, L.M., and Johnson, W. (1992) ``Case-Deletion Diagnostics for Mixed Models, Technometrics, 34, 38 -- 45.

Dillane, D. (2005), Deletion Diagnostics for the Linear Mixed Model,'' Ph.D. thesis, Trinity College Dublin.

Schabenberger, O. (2004) Mixed Model Influence Diagnostics, in Proceedings of the Twenty-Ninth SAS Users Group International Conference, SAS Users Group International.

See also

Examples

data(sleepstudy, package = 'lme4')
fm <- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# Subject level deletion and diagnostics
subject.del  <- case_delete(model = fm, group = "Subject", type = "both")

#> Warning: group is not a valid argument for this function. As of version 0.4.0, group has been replaced by level. See ?hlm_influence for more information.

#> Warning: Model failed to converge with max|grad| = 0.0027157 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.0021305 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00239345 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00237651 (tol = 0.002, component 1)

subject.diag <- diagnostics(subject.del)

#> Warning: 'diagnostics' is deprecated.
#> Use 'hlm_influence' instead.
#> See help("Deprecated")