Wages for male high school dropouts

Data on the labor-market experience of male high school dropouts.

Format

A data frame with 6402 observations on the following 15 variables.

id

respondent id - a factor with 888 levels.

lnw

natural log of wages expressed in 1990 dollars.

exper

years of experience in the work force

ged

equals 1 if respondent has obtained a GED as of the time of survey, 0 otherwise

postexp

labor force participation since obtaining a GED (in years) - before a GED is earned postexp = 0, and on the day a GED is earned postexp = 0

black

factor - equals 1 if subject is black, 0 otherwise

hispanic

factor - equals 1 if subject is hispanic, 0 otherwise

hgc

highest grade completed - takes integers 6 through 12

hgc.9

hgc - 9, a centered version of hgc

uerate

local area unemployment rate for that year

ue.7

ue.centert1

ue.mean

ue.person.cen

ue1

Source

These data are originally from the 1979 National Longitudinal Survey on Youth (NLSY79) that can be found here https://www.bls.gov/nls/nlsy79.htm.

Singer and Willett (2003) used these data for examples in chapter (insert info. here) and the data sets used can be found on the UCLA Statistical Computing website: https://stats.idre.ucla.edu/other/examples/alda/

Additionally the data were discussed by Cook and Swayne (2003) and the data can be found on the GGobi website: http://ggobi.org/book.html.

References

Singer, J. D. and Willett, J. B. (2003), Applied Longitudinal Data Analysis: Modeling Change and Event Occurrence, New York: Oxford University Press.

Cook, D. and Swayne, D. F. (2007), Interactive and Dynamic Graphics for Data Analysis with R and GGobi, Springer.

Examples

str(wages)

#> 'data.frame':	6402 obs. of  15 variables:
#>  $ id           : Factor w/ 888 levels "31","36","53",..: 1 1 1 1 1 1 1 1 2 2 ...
#>  $ lnw          : num  1.49 1.43 1.47 1.75 1.93 ...
#>  $ exper        : num  0.015 0.715 1.734 2.773 3.927 ...
#>  $ ged          : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ postexp      : num  0.015 0.715 1.734 2.773 3.927 ...
#>  $ black        : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ hispanic     : Factor w/ 2 levels "0","1": 2 2 2 2 2 2 2 2 1 1 ...
#>  $ hgc          : int  8 8 8 8 8 8 8 8 9 9 ...
#>  $ hgc.9        : int  -1 -1 -1 -1 -1 -1 -1 -1 0 0 ...
#>  $ uerate       : num  3.21 3.21 3.21 3.3 2.89 2.49 2.6 4.8 4.89 7.4 ...
#>  $ ue.7         : num  -3.79 -3.79 -3.79 -3.71 -4.11 ...
#>  $ ue.centert1  : num  0 0 0 0.08 -0.32 ...
#>  $ ue.mean      : num  3.21 3.21 3.21 3.21 3.21 3.21 3.21 3.21 5.1 5.1 ...
#>  $ ue.person.cen: num  0 0 0 0.08 -0.32 ...
#>  $ ue1          : num  3.21 3.21 3.21 3.21 3.21 3.21 3.21 3.21 4.89 4.89 ...

summary(wages)

#>        id            lnw            exper             ged        
#>  1204   :  13   Min.   :0.708   Min.   : 0.001   Min.   :0.0000  
#>  3440   :  13   1st Qu.:1.591   1st Qu.: 1.609   1st Qu.:0.0000  
#>  7373   :  13   Median :1.842   Median : 3.451   Median :0.0000  
#>  9968   :  13   Mean   :1.897   Mean   : 3.957   Mean   :0.2719  
#>  10392  :  13   3rd Qu.:2.140   3rd Qu.: 5.949   3rd Qu.:1.0000  
#>  12043  :  13   Max.   :4.304   Max.   :12.700   Max.   :1.0000  
#>  (Other):6324                                                    
#>     postexp        black    hispanic      hgc             hgc.9         
#>  Min.   : 0.0000   0:4783   0:4859   Min.   : 6.000   Min.   :-3.00000  
#>  1st Qu.: 0.0000   1:1619   1:1543   1st Qu.: 8.000   1st Qu.:-1.00000  
#>  Median : 0.0000                     Median : 9.000   Median : 0.00000  
#>  Mean   : 0.9076                     Mean   : 8.948   Mean   :-0.05248  
#>  3rd Qu.: 0.1168                     3rd Qu.:10.000   3rd Qu.: 1.00000  
#>  Max.   :12.2600                     Max.   :12.000   Max.   : 3.00000  
#>                                                                         
#>      uerate            ue.7          ue.centert1         ue.mean      
#>  Min.   : 1.790   Min.   :-5.2050   Min.   :-15.310   Min.   : 2.890  
#>  1st Qu.: 5.390   1st Qu.:-1.6050   1st Qu.: -2.900   1st Qu.: 6.070  
#>  Median : 7.000   Median : 0.0000   Median : -0.500   Median : 7.150  
#>  Mean   : 7.733   Mean   : 0.7297   Mean   : -1.015   Mean   : 7.734  
#>  3rd Qu.: 9.400   3rd Qu.: 2.3000   3rd Qu.:  0.600   3rd Qu.: 8.650  
#>  Max.   :23.700   Max.   :16.7050   Max.   : 13.005   Max.   :17.130  
#>                   NA's   :402       NA's   :406                       
#>  ue.person.cen            ue1        
#>  Min.   :-8.403000   Min.   : 2.890  
#>  1st Qu.:-1.350500   1st Qu.: 6.200  
#>  Median :-0.142000   Median : 8.300  
#>  Mean   :-0.000011   Mean   : 8.762  
#>  3rd Qu.: 1.133000   3rd Qu.:10.200  
#>  Max.   :13.421000   Max.   :23.700  
#>                                      

if (FALSE) {
library(lme4)
lmer(lnw ~ exper + (exper | id), data = wages)
}